tudo double e tao as 3 a dar penso

+++ UMeR/Cliente.java
@@ -3,13 +3,13 @@ import java.time.LocalDate;
 
 public class Cliente extends Actor
 {
-    private float totalGasto = 0;
+    private Double totalGasto = 0.0;
 
     public Cliente(){
         super();  
     }
    
-    public float getTotalGasto(){
+    public Double getTotalGasto(){
         return this.totalGasto;
     }
 
@@ -28,10 +28,10 @@ public class Cliente extends Actor
         Motorista m = new Motorista("ola","ola","ola","ola",da);
         Veiculo v = new Veiculo("ola",2,3,x);
         
-        Viagem via = new Viagem(1,this,x,y,10,2,3,m,v,da);
-        Viagem via2 = new Viagem(2,this,x,y,10,2,3,m,v,db);
-        Viagem via3 = new Viagem(3,this,x,y,10,2,3,m,v,dc);
-        Viagem via4 = new Viagem(3,this,x,y,10,2,3,m,v,di);
+        Viagem via = new Viagem(1,this,x,y,10,2,3.0,m,v,da);
+        Viagem via2 = new Viagem(2,this,x,y,10,2,3.0,m,v,db);
+        Viagem via3 = new Viagem(3,this,x,y,10,2,3.0,m,v,dc);
+        Viagem via4 = new Viagem(3,this,x,y,10,2,3.0,m,v,di);
         listaViagens.add(via);
         listaViagens.add(via2);
         listaViagens.add(via3);
